Posthtml templating

+++ src/generators/output/toDisk.js
@@ -3,39 +3,46 @@ const fs = require('fs-extra')
 const fm = require('front-matter')
 const glob = require('glob-promise')
 const deepmerge = require('deepmerge')
-const stripHTML = require('string-strip-html')
-const { asyncForEach } = require('../../utils/helpers')
+const removePlaintextTags = require('../../transformers/plaintext')
+const { asyncForEach, getPropValue } = require('../../utils/helpers')
 
 const Config = require('../config')
 const Tailwind = require('../tailwind')
+const Plaintext = require('../plaintext')
 
 const render = require('./toString')
 
 module.exports = async (env, spinner) => {
-  const config = await Config.getMerged(env).catch(err => { spinner.fail('Build failed'); console.log(err); process.exit() })
-  const css = await Tailwind.fromFile(config, env).catch(err => { spinner.fail('Build failed'); console.log(err); process.exit() })
+  const config = await Config.getMerged(env).catch(err => { spinner.fail('Build failed'); console.log(err); process.exit(1) })
+  const css = await Tailwind.fromFile(config, env).catch(err => { spinner.fail('Build failed'); console.log(err); process.exit(1) })
 
-  const sourceDir = config.build.templates.source
-  const outputDir = config.build.destination.path
+  const sourceDir = getPropValue(config, 'build.templates.root') || 'src/templates'
+  const outputDir = getPropValue(config, 'build.destination.path') || `build_${env}`
+  let filetypes = getPropValue(config, 'build.templates.extensions') || 'html'
 
-  await fs.remove(outputDir)
-
-  if (Array.isArray(sourceDir)) {
-    await asyncForEach(sourceDir, path => fs.copy(path, outputDir))
-  } else {
-    await fs.copy(sourceDir, outputDir)
+  if (fs.pathExistsSync(outputDir)) {
+    await fs.remove(outputDir)
   }
-
-  let filetypes = config.build.templates.filetypes || 'html|njk|nunjucks'
 
   if (Array.isArray(filetypes)) {
     filetypes = filetypes.join('|')
   }
 
+  if (Array.isArray(sourceDir)) {
+    await asyncForEach(sourceDir, async source => {
+      await fs.copy(source, outputDir).catch(error => spinner.warn(error.message))
+    })
+  } else {
+    await fs.copy(sourceDir, outputDir).catch(error => spinner.warn(error.message))
+  }
+
   const templates = await glob(`${outputDir}/**/*.+(${filetypes})`)
 
   if (templates.length < 1) {
-    throw RangeError(`No "${filetypes}" templates found in \`${sourceDir}\`. If the path is correct, please check your \`build.templates.filetypes\` config setting.`)
+    spinner
+      .fail(`Error: no files with the .${filetypes} extension found in your \`templates.root\` path${Array.isArray(sourceDir) ? 's' : ''}`)
+      .fail('Build failed')
+    process.exit(1)
   }
 
   if (config.events && typeof config.events.beforeCreate === 'function') {
@@ -47,51 +54,60 @@ module.exports = async (env, spinner) => {
     const frontMatter = fm(html)
     const templateConfig = deepmerge(config, frontMatter.attributes)
     const events = templateConfig.events || []
+
     templateConfig.isMerged = true
+    templateConfig.env = env
 
-    html = await render(html, {
-      tailwind: {
-        compiled: css
-      },
-      maizzle: {
-        config: templateConfig
-      },
-      env: env,
-      ...events
-    })
+    try {
+      html = await render(html, {
+        tailwind: {
+          compiled: css
+        },
+        maizzle: {
+          config: templateConfig
+        },
+        ...events
+      })
+    } catch (error) {
+      switch (templateConfig.build.fail) {
+        case 'silent':
+          spinner.warn(`Failed to compile ${file}`)
+          break
+        case 'verbose':
+          spinner.warn(`Failed to compile ${file}`)
+          console.error(error)
+          break
+        default:
+          spinner.fail(`Failed to compile ${file}`)
+          throw error
+      }
+    }
 
-    const ext = templateConfig.build.destination.extension || 'html'
+    if (templateConfig.plaintext) {
+      await Plaintext.output(html, file, templateConfig)
+    }
+
+    html = removePlaintextTags(html, config)
 
     fs.outputFile(file, html)
-      .then(() => {
-        if (templateConfig.plaintext) {
-          const plaintext = stripHTML(html,
-            {
-              dumpLinkHrefsNearby: {
-                enabled: true,
-                putOnNewLine: true,
-                wrapHeads: '[',
-                wrapTails: ']'
-              }
-            })
-
-          const filepath = templateConfig.permalink || file
-          const plaintextPath = path.join(path.dirname(filepath), path.basename(filepath, path.extname(filepath)) + '.txt')
-
-          fs.outputFileSync(plaintextPath, plaintext)
-        }
-
+      .then(async () => {
         if (templateConfig.permalink) {
-          return fs.move(file, templateConfig.permalink, { overwrite: true })
+          await fs.move(file, templateConfig.permalink, { overwrite: true })
         }
 
-        const parts = path.parse(file)
-        fs.rename(file, `${parts.dir}/${parts.name}.${ext}`)
+        const extension = getPropValue(templateConfig, 'build.destination.extension') || 'html'
+
+        if (extension !== 'html') {
+          const parts = path.parse(file)
+          await fs.rename(file, `${parts.dir}/${parts.name}.${extension}`)
+        }
       })
   })
 
-  if (fs.pathExistsSync(config.build.assets.source)) {
-    await fs.copy(config.build.assets.source, `${outputDir}/${config.build.assets.destination}`)
+  const assets = getPropValue(config, 'build.assets') || { source: '', destination: 'assets' }
+
+  if (fs.pathExistsSync(assets.source)) {
+    await fs.copy(assets.source, `${outputDir}/${assets.destination}`)
   }
 
   if (config.events && typeof config.events.afterBuild === 'function') {
